Update kotest and other minor parts
