+++ arrow-check-kotest/build.gradle
@@ -9,7 +9,6 @@ apply from: "$DOC_CREATION"
 dependencies {
     api project(":arrow-check")
 
-    api("io.kotlintest:kotlintest-runner-junit5:3.4.2") {
-        exclude group: "io.arrow-kt"
-    }
+    api "io.kotest:kotest-runner-junit5-jvm:$KOTEST_VERSION" // for kotest framework
+    api "io.kotest:kotest-assertions-core-jvm:$KOTEST_VERSION" // for kotest core jvm assertions
 }
