+++ arrow-check-kotest/gradle.properties
@@ -1,0 +1,4 @@
+# Maven publishing configuration
+POM_NAME=Arrow-Check-Kotest
+POM_ARTIFACT_ID=arrow-check-kotest
+POM_PACKAGING=jar
