--- arrow-check-kotlintest/gradle.properties
@@ -1,4 +1,0 @@
-# Maven publishing configuration
-POM_NAME=Arrow-Check-Kotlintest
-POM_ARTIFACT_ID=arrow-check-kotlintest
-POM_PACKAGING=jar
