+++ settings.gradle
@@ -1,3 +1,3 @@
 rootProject.name = 'arrow-check-lib'
 
-include 'arrow-check', 'arrow-check-kotlintest'
+include 'arrow-check', 'arrow-check-kotest'
